requirements.txt
python-telegram-bot
pandas
openpyxl
python-dotenv
camelot-py[cv]
tabula-py
ghostscript
Pillow